pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
    "tzdata>=2025.2",
    "SQLAlchemy>=2.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "jinja2>=3.1",
]
